+++ generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.5.0
+
 ## 0.4.3
 
 * Minor fix